pyproject.toml
[project]
name = "durable-agentic-quantum"
version = "0.0.1"
description = "Resilient Qiskit Circuit Generation"
readme = "README.md"
requires-python = ">=3.13"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "fastmcp==2.13.3",
    "google-genai==1.56.0",
    "matplotlib==3.10.8",
    "mcp==1.22.0",
    "pylatexenc==2.10",
    "python-fasthtml==0.12.36",
    "qiskit==2.2.3",
    "temporalio==1.21.1",
]

[dependency-groups]
dev = [
    "pyrefly==0.46.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "ruff==0.14.10",
]

[tool.pyrefly]
search_path = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/app"]
omit = ["*__init__.py"]
